refactor(mobile): modularize encryption service into focused, maintainable modules
