+++ apps/mobile/v1/src/lib/encryption/storage/secureStorage.ts
@@ -1,0 +1,123 @@
+/**
+ * Secure Storage Wrapper
+ * Abstraction layer over expo-secure-store
+ */
+
+import * as SecureStore from 'expo-secure-store';
+import * as Crypto from 'expo-crypto';
+import { STORAGE_KEYS } from './storageKeys';
+import { arrayBufferToBase64 } from '../core/crypto';
+
+/**
+ * User secret management - in-memory cache
+ */
+const userSecretsCache = new Map<string, string>();
+
+/**
+ * Get or generate a user-specific secret
+ */
+export async function getUserSecret(userId: string): Promise<string> {
+  // Check master key first
+  try {
+    const masterKey = await SecureStore.getItemAsync(STORAGE_KEYS.MASTER_KEY(userId));
+    if (masterKey) {
+      return masterKey;
+    }
+  } catch {
+    // Continue to fallback
+  }
+
+  // Check memory cache
+  if (userSecretsCache.has(userId)) {
+    return userSecretsCache.get(userId)!;
+  }
+
+  // Use secure storage for user secrets
+  const storageKey = STORAGE_KEYS.USER_SECRET(userId);
+  try {
+    let secret = await SecureStore.getItemAsync(storageKey);
+
+    if (!secret) {
+      // Generate new random secret and store it securely
+      const randomBytes = await Crypto.getRandomBytesAsync(64);
+      secret = arrayBufferToBase64(randomBytes);
+      await SecureStore.setItemAsync(storageKey, secret);
+    }
+
+    // Cache in memory for performance
+    userSecretsCache.set(userId, secret);
+    return secret;
+  } catch (error) {
+    throw new Error(`Failed to get user secret: ${error}`);
+  }
+}
+
+/**
+ * Store master key in secure storage
+ */
+export async function storeMasterKey(userId: string, keyString: string): Promise<void> {
+  await SecureStore.setItemAsync(STORAGE_KEYS.MASTER_KEY(userId), keyString);
+  await SecureStore.setItemAsync(STORAGE_KEYS.HAS_MASTER_PASSWORD(userId), 'true');
+}
+
+/**
+ * Get master key from secure storage
+ */
+export async function getMasterKey(userId: string): Promise<string | null> {
+  try {
+    return await SecureStore.getItemAsync(STORAGE_KEYS.MASTER_KEY(userId));
+  } catch {
+    return null;
+  }
+}
+
+/**
+ * Check if user has master password set
+ */
+export async function hasMasterPasswordFlag(userId: string): Promise<boolean> {
+  try {
+    const flag = await SecureStore.getItemAsync(STORAGE_KEYS.HAS_MASTER_PASSWORD(userId));
+    return flag === 'true';
+  } catch {
+    return false;
+  }
+}
+
+/**
+ * Clear all encryption data for a user
+ */
+export async function clearUserStorageData(userId: string): Promise<void> {
+  const keysToDelete = [
+    STORAGE_KEYS.MASTER_KEY(userId),
+    STORAGE_KEYS.HAS_MASTER_PASSWORD(userId),
+    STORAGE_KEYS.TEST_ENCRYPTION(userId),
+    STORAGE_KEYS.USER_SECRET(userId),
+  ];
+
+  for (const key of keysToDelete) {
+    try {
+      await SecureStore.deleteItemAsync(key);
+      if (__DEV__) {
+        console.log(`✅ Deleted ${key}`);
+      }
+    } catch (error) {
+      if (__DEV__) {
+        console.log(`❌ Failed to delete ${key}:`, error);
+      }
+    }
+  }
+
+  // Clear from memory cache
+  userSecretsCache.delete(userId);
+}
+
+/**
+ * Delete old user secret key
+ */
+export async function deleteOldUserSecret(userId: string): Promise<void> {
+  try {
+    await SecureStore.deleteItemAsync(STORAGE_KEYS.USER_SECRET(userId));
+  } catch {
+    // Ignore if doesn't exist
+  }
+}
